Take CAPTURE_FAILED, REFUNDED_REVERSED and REFUND_FAILED into account
